remove unused functions
